update:マップページでSSRするため、mapをコンポーネント化しました
